Makes acceptance testing modular
